+++ src/tableau_api_lib/utils/querying/groups.py
@@ -1,67 +1,71 @@
-"""
-Helper functions for querying REST API data for groups
-"""
+"""This module defines helper functions for querying REST API data for groups."""
 
+from typing import Any, Dict, List, Optional
 
 import pandas as pd
+from typeguard import typechecked
+
 from tableau_api_lib import TableauServerConnection
 from tableau_api_lib.utils import extract_pages
 from tableau_api_lib.utils.querying import get_users_dataframe
 
 
-def get_all_group_fields(conn):
-    all_groups = extract_pages(conn.query_groups, parameter_dict={'fields': 'fields=_all_'})
+@typechecked
+def get_all_group_fields(conn: TableauServerConnection) -> List[Dict[str, Any]]:
+    """Returns details for all groups in the Tableau Server environment, including all queryable fields."""
+    all_groups = extract_pages(conn.query_groups, parameter_dict={"fields": "fields=_all_"})
     return all_groups
 
 
-def get_group_users(conn, group_id):
-    all_group_users = extract_pages(conn.get_users_in_group,
-                                    content_id=group_id,
-                                    parameter_dict={'fields': 'fields=_all_'})
+@typechecked
+def get_group_users(conn: TableauServerConnection, group_id: str) -> Dict[str, Any]:
+    """Returns details of users belonging to the specified Tableau group."""
+    all_group_users = extract_pages(
+        conn.get_users_in_group, content_id=group_id, parameter_dict={"fields": "fields=_all_"}
+    )
     return all_group_users
 
 
-def get_all_group_names(conn):
+@typechecked
+def get_all_group_names(conn: TableauServerConnection) -> List[str]:
+    """Returns a list of all groups available in the Tableau Server environment."""
     all_groups = get_all_group_fields(conn)
-    all_groupnames = [group['name'] for group in all_groups]
+    all_groupnames = [group["name"] for group in all_groups]
     return all_groupnames
 
 
-def get_all_group_domain_names(conn):
+@typechecked
+def get_all_group_domain_names(conn: TableauServerConnection) -> List[str]:
+    """Returns a list of domain names for all groups in the Tableau Server environment."""
     all_groups = get_all_group_fields(conn)
-    all_group_roles = [group['domain']['name'] for group in all_groups]
+    all_group_roles = [group["domain"]["name"] for group in all_groups]
     return all_group_roles
 
 
-def get_groups_dataframe(conn) -> pd.DataFrame:
-    """
-    Get a Pandas DataFrame containing details for all groups on the active site.
-    :param TableauServerConnection conn: the Tableau Server connection
-    :return: pd.DataFrame
-    """
+@typechecked
+def get_groups_dataframe(conn: TableauServerConnection) -> pd.DataFrame:
+    """Returns a Pandas DataFrame containing details for all groups on the active site."""
     groups_df = pd.DataFrame(get_all_group_fields(conn))
-    groups_df['domain'] = groups_df['domain'].apply(lambda x: x['name'])
+    groups_df["domain"] = groups_df["domain"].apply(lambda x: x["name"])
     return groups_df
 
 
-def get_group_users_dataframe(conn, group_id) -> pd.DataFrame:
-    """
-    Gets a Pandas DataFrame containing all users belonging to the specified group.
-    :param TableauServerConnection conn: the Tableau Server connection
-    :param str group_id: the ID for the group whose users are being queried
-    :return: pd.DataFrame
-    """
+@typechecked
+def get_group_users_dataframe(conn: TableauServerConnection, group_id: str) -> pd.DataFrame:
+    """Returns a Pandas DataFrame describing all users belonging to the specified Tableau group."""
     group_users_df = pd.DataFrame(get_group_users(conn, group_id))
     users_df = get_users_dataframe(conn)
     users_columns = list(users_df.columns)
-    all_group_users = group_users_df.merge(users_df,
-                                           how='left',
-                                           left_on='id',
-                                           right_on='id',
-                                           suffixes=('delete', None))
-    return all_group_users[users_columns]
+    if group_users_df.empty is True:
+        return group_users_df
+    else:
+        all_group_users_df = group_users_df.merge(
+            users_df, how="left", left_on="id", right_on="id", suffixes=("delete", None)
+        )
+        return all_group_users_df[users_columns]
 
 
+@typechecked
 def get_groups_for_a_user_dataframe(conn: TableauServerConnection, user_id: str) -> pd.DataFrame:
     """Returns a Pandas DataFrame containing all groups a user belongs to."""
     groups_for_user_df = pd.DataFrame(extract_pages(conn.get_groups_for_a_user, content_id=user_id))